Add configurable API URL via properties file
Moved the API URL configuration from a hardcoded value to a property loaded from 'config/rankeduta.properties'. Added logic to load and store the API URL, and to warn and stop the server if it is not set. Updated HTTPClient to use the new configurable API URL. Also changed SendStorageHandler to send storage data as a string.